Update csp-directives to be more testable
