+++ config/development.json
@@ -3,6 +3,7 @@
     "base": "http://localhost:3000"
   },
   "features": {
+    "enableAllowLocalhost": false,
     "enableAwardsForAllApplications": true,
     "enableSeeders": true,
     "enableSalesforceConnector": true
