+++ config/test.json
@@ -1,5 +1,6 @@
 {
   "features": {
+    "enableAllowLocalhost": false,
     "enableAwardsForAllApplications": true,
     "enableDigitalFundApplications": false,
     "enableHotjar": true,
